Fix: BigInt error, half-donut orientation, warning card accent, white sidebar
- Fix BigInt TypeError: cast row_count::int in SQL so JS receives a plain number
- HalfDonut: rewrite using circle + strokeDasharray + rotate(180) so the arc
  correctly draws the TOP semicircle (9 o-clock → 12 → 3 o-clock)
- DomainCard: add colored top-border accent on cards with warnings
  (border-t-red-400 for ≥3 alerts, border-t-amber-400 for 1-2 alerts)
  in addition to the triangle icon, making warning visible at card level
- Sidebar: white background + brand-blue text to match colour scheme;
  active item uses bg-brand-purple/10 with brand-deep text and left accent bar;
  hover uses canvas background with brand-deep text

+++ lib/queries/dashboard.ts
@@ -36,7 +36,7 @@ export async function getRecentAssets(userId: string, limit = 6): Promise<Recent
       asset_id    AS "assetId",
       asset_name  AS "assetName",
       asset_meta  AS "assetMeta",
-      row_count   AS "rowCount",
+      row_count::int   AS "rowCount",
       visited_at  AS "visitedAt"
     FROM bayanat.user_recent_assets
     WHERE user_id = ${userId}
